feat: wip 1.26.30 support

+++ src/main/java/cn/nukkit/inventory/transaction/data/UseItemData.java
@@ -21,5 +21,5 @@ public class UseItemData implements TransactionData {
     public Vector3f clickPos;
     public int blockRuntimeId;
     public int clientInteractPrediction;
-    public byte clientCooldownState;
+    public int clientCooldownState;
 }
